GPU ppm 3d (GPU hydro PR 2/n) (#77)
This offloads ppm_2d and mkutrans_2d to the GPU for ppm types 1 and 2. It has been tested for the 2d reacting bubble problem compiled for MPI and GPU.

Note: this version will not work for ppm type 0, as this is not yet implemented (see #78). It is also likely that it shall not work if compiled with OpenMP (this will come in a future PR)

This offloads mkutrans_3d and ppm_3d to the GPU for ppm types 1 and 2.It has been tested for the 3d reacting bubble & wdconvect problems compiled for MPI and GPU.

Note: this version will not work for ppm type 0, as this is not yet implemented (see #78). It is also likely that it shall not work if compiled with OpenMP (this will come in a future PR)