Add diagnostics for planar geometries (#90)
Diagnostics currently are spherical only, this PR extends the same measurements to a planar geometry. In 2D, things like z-components are just recorded as zero, and spherical-specific things such as measurements at the star centre are not recorded.

The fortran now has separate diag and diag_sphr subroutines (where diag_sphr is what was previously diag). They basically do the same thing in this PR. Gravitational potential energy is defined differently so it too has separate planar and spherical subroutines.

* add diagnostics for planar mode

* tidy comments

* remove 1d support

* rename variables from coordenuc2 -> coord_enuc_y etc

* remove remenants of a bad copy paste

* T_max -> temp_max, coord_Tmax -> coord_temp_max, etc, in the fortran

* bug fix - coord_enuc was being set to coord_tmax

* remove the rest of the unnecessary AMREX_SPACEDIM < 2 conditionals (remove 1d support)

* move module <name> to 1st line to make work with sphinx (I think this is all thats required?

* guard star center calculations with preprocessor directive

* clarify comments that will end up in docs

* add a rudimentary test problem for planar diagnostics

* fix a typo in the python notebook

* remove hard-coded dimensional factor

* delete print statement / unnecessary print statement