Convert MakeEdgeScal Fortran kernels to C++ lambdas (#133)
This converts the Fortran kernel launches in the MakeEdgeScal routine to C++ lambda functions. The performance on CPU is essentially unchanged, but on GPU I see a speed up of 20-30x for this function, leading to an speed up overall of ~30% for Maestro::Evolve.

The actual code itself is essentially unchanged - I've just added punctuation and fixed the array indexing to match C++ counting from 0.

I've tested this for 3d reacting_bubble and wdconvect and the answers seem to match development.

Conversion of the 2d routines and slope routines (for ppm type 0) will come in later PRs.