Convert VelPred routines to C++ (#136)
This converts the velocity prediction routines to C++ lambda functions.

Testing this for reacting_bubble with the burning switched off, there are some reasonably large differences for all routines apart from VelPredInterface. I can't see any obvious errors so this is hopefully just due to roundoff error, but it could be that there are one or more bugs lurking that I haven't spotted.

It compiles and runs on the GPU without dying.