Convert vorticity routines to C++  (#139)
This converts to vorticity routines to C++. These routines are called to calculate the vorticity when generating plotfiles, and were not offloaded to GPU because they included Fortran functions defined within a subroutine. Converting them to C++ lambda functions therefore significantly speeds up the generation of plotfile data for GPU runs.