this merges the _parameters into a single module (#146)
This merges the _parameters into a single module previously we had two sets of parameters, some lived in probin_module and some in extern_probin_module, but they were initialized with the same namelist.  This made the script more complicated to maintain.  Now all the parameters live in probin_module and a stub extern_probin_module is created that just `uses` probin_module to ensure the Microphysics works.