Convert advect_base.F90 routines to C++ (#148)
This converts the routines in advect_base.F90 and make_edge_scal.F90 to C++. It also converts a couple of the routines in base_state_geometry.F90 in order to do base state restriction and ghost filling.

The indexing of the 1d base states is quite messy at the moment, but should become clearer when we switch to storing the base state as a Vector of RealVectors.

This should not change the results of single level problems. Testing 3d reacting_bubble with 3 levels gives no changes for ppm_type=0, and small round-off level changes for types 1 and 2. It compiles and runs on the GPU.

* converting make_edge_state

* fixed compilation

* started converting make_edge_state_1d_planar

* (mostly) converted advect_base

* it compiles

* it works for spherical=1, ppm_type=0

* it works for spherical=1, ppm_type=0

* it works for spherical=1, ppm_type=1

* it works for spherical=1, ppm_type=2

* it works for spherical=0, ppm_type=0

* it works for spherical=0, ppm_type=1

* it works for spherical=0, ppm_type=2

* RestructBase and FillGhostBase seem to work with amr

* gives no diffs for AMR for ppm_type=0, and round-off diffs for ppm_type=1 or 2

* removed fortran files

* GPU fixes

* fixed some loops to reduce errors in multilevel wdconvect

* max_radial_level -> finest_radial_level

Co-authored-by: Adam Jacobs <ajacobs@pa.msu.edu>