Convert base_state.F90 to C++ (#167)
This converts the init_base_state routine to C++, both in the Src_nd directory and for all the custom base state initialization functions in the problem directories.

It also converts some of the functions in the Util directory (model_parser and simple_log, and adds another util directory, the purpose of which is to contain general small utility functions (e.g. so far it contains a function for trimming C++ strings of whitespace).

This PR depends on #165, and is untested except for rt.