Convert debug. diag and tagging routines to C++ (#173)
This converts the debug, diagnostics and tagging routines to C++. It does not depend on any previous PRs and should not change the output.

Note that by default all routines will now use the tagging routines defined in MaestroTagging.cpp, and they do not need to create a local copy for tagging to work. A local copy is now only needed if problem-specific customizations are required.