Add Aux data quantities (#211)
This adds the auxiliary data (addressing #202). In this PR, the quantities are initialized in the initial data for the reacting bubble test problem, however in a subsequent PR this shall be moved to a separate post-initial data step.

I also found a bug in the RhoXFlux routine that didn't look thread safe, so moved the density reset outside of the loops over the components.