Make GPU test runs deterministic (#229)
This adds in a few changes to make GPU test runs (i.e. GPU runs with the flag maestro.deterministic_nodal_solve=true set) deterministic (or at least agree to within a small enough error that the test suite will consider their plotfiles to agree).

The main changes are switching min and max to always use the amrex:: versions, adding a few Gpu::synchronize(); statements (which should have been there in the first place) and switching off the kernel launch region in a couple of places so that those regions always run on the CPU for tests.

I've tested this for RT (2d and 3d) and wdconvect.