Initialize density from base state for Urca (#323)
The current method initializes the density by summing the partial densities (rho X) from the composition.

The Urca initial models have steep compositional changes which lead to minor interpolation errors because of limiting. These are negligible as far as the reactions go, but because density is derived from this composition the simulation begins with some small density perturbations. These go on to grow larger and disrupt the simulation fairly quickly. see #321

Instead by initializing the density directly from the base state, we ensure a smooth density profile.