check the burner status and abort if we fail (#379)
this shifts the responsibility for aborting from Microphysics to MAESTROeX.
Previously we would not abort on GPUs.  Now we do, since we do a reduce_op
and handle the abort outside of the lambda kernel.

Note: the SDC burner is not changed, since the SDC code path does not currently compile.